Only bulk fill once
Ideally, re-attempting bulk fill a second time would try to place all
the same items at the same locations they could not be placed the last
time. In practice, because items are picked 1 player at a time, this
usually doesn't happen because the number of unplaceable items per
player in the first attempt will vary. Either way, additional attempts
with the same locations are likely to result in the same locations being
unreachable, so there is not much reason to re-attempt a bulk fill.