Core: Cleanup + Warn for single_player_placement fill with multiple players
#2415 removed the ability to provide items/locations for multiple
players simultaneously with single_player_placement=True. Now that this
is removed, the old behaviour should be considered deprecated so that
developers using the old behaviour don't run into unexpected problems.

With the requirement in place that all items in the provided item pool
have to belong to the same player, that singular player can be retrieved
once in advance rather than getting the player from the most recent item
set into the `item` variable. This reduces the used scope of the `item`
variable to just the loops it is part of and removes IDE warnings about
the variable possibly not being initialized.