Cull as many items as possible, like the current implementation
Allow required locations to be collected in later spheres

Re-arrange spheres when culling locations
Updating the spheres for the new sphere of each location makes later
iterations faster by not having to carry along so many items between
spheres.

A bit slower, but still a good improvement over the implementation in
this main branch.

Given the size of the culling code now, it should probably be
deduplicated between the sphere culling and precollected culling.