Add support for entrance rules depending on other worlds
Previously, region accessibility would only be staled when and item
belonging to the same world was collected into the state.

This meant that if a world added entrance rules depending on another
world then the item required to access that entrance could be collected,
but the region accessibility cache could never be staled, leading to the
entrance never being checked.

CollectionState.collect() and CollectionState.remove() now find all
players logically dependent on the owning player of the
collected/removed item and stale the cache for each of those players.