SM: Speed up deepcopy in copy_mixin (#4228)
SMBoolManager did not implement __deepcopy__, so using copy.deepcopy in
the copy_mixin was slow and possibly incorrect.

Best I can tell, copy.deepcopy would have been effectively pickling and
then unpickling using object.__reduce__ and object.__setstate__
recursively for the SMBoolManager and the contents of it's __dict__ for
everything that did not implement __deepcopy__.

Notably, the new __deepcopy__ prevents the Objectives instance being
deepcopied. Objectives instances are cached per-player in
Objectives.objDict, so it was resulting in a lot of wasteful copying.
The deepcopied Objectives instance would also not end up in the
per-player cache, which did not seem intended.

Generating with 10 copies of the template Super Metroid yaml with
progression balancing disabled and no output, generation goes from about
6.9s to about 2.2s for me.

This patch is about the best as I can understand from the randomizer
code. It is entirely possible that further performance improvements
could be made by someone more familiar with the code.