Revert loading all worlds from settings._update_cache()
Trying to load all worlds here would load all worlds even when trying to
generate with only a few worlds.

The late loading of these worlds would result in worlds adding copy/init
mixins after CollectionState instances have already been created,
meaning that copying any existing CollectionState would fail by trying
to copy attributes the new copy mixins were expecting to exist, but did
not because the CollectionState was created before the init mixin was
added.

This probably means it would be a good idea for there to be a function
that can be called to prevent further world loading. Possibly raising an
error if such loading is attempted.