Merge PR 3812
What is this fixing or adding?

Increases the performance of sweeping state in CollectionState.sweep_for_events() and MultiWorld.can_beat_game, without changing the generated output (at least for supported games with their template yamls).

CollectionState.sweep_for_events() takes up a significant amount of generation time, mostly from filling progression items, and the majority when not generating a playthrough, so improvements to its performance can have a noticeable effect on generation duration.

Similarly, MultiWorld.can_beat_game() takes up a significant amount of playthrough calculation time from culling spheres.

This patch reduces the duration of both and refactors them to use the same sweeping code.

This patch has been split from #3688 to focus only on the sweep itself, and has been simplified to not require any additions to the existing world API.
Make better use of region accessibility caching

Region accessibility is cached per-player and the cache for that player becomes stale whenever an item is collected or removed for that player. The original sweeping implementations make use of this by only collecting items once all reachable locations have been found for all players.

This patch changes to the sweep to find reachable locations per-player, collect the items from those locations and then move onto the next player. This keeps region accessibility cached as before, but has the advantage that if a player's reachable locations contain an item that gives a later player access to more locations, it will take fewer iterations for all reachable locations to be collected from.
Skip checking location accessibility for locations belonging to a player which did not receive any advancement in the previous iteration

If no advancement items belonging to a player were collected in the previous sweep iteration (after the point at which their reachable locations were found), then it is assumed that player should have zero newly reachable locations, so checking for their reachable locations can be skipped.

While almost all worlds only logically depend on only their own items/locations etc., it is possible for a world to logically depend on a part of another world. To account for this, once the sweep runs out of accessible locations, the locations of all players are checked to confirm that the sweep is indeed finished.

My previous idea to account for worlds logically depending on other worlds was to require all worlds to specify which worlds they logically depended on. Performing extra sweeps to confirm will be slower than this previous idea, but it is simpler and does not require any changes to the existing world API.
How was this tested?

Fixed seeds were generated, with and without this patch, with different sets of template yamls in the Players directory, using games known to have deterministic generation. The output from each set of template yamls was seen to match. Playthrough calculation is non-deterministic, so was not considered here.

Seeds using item links have also been tested.

To test the case of worlds depending on other worlds, I modified the A Hat in Time world to find all other AHiT worlds in the multiworld and then added additional rules to a few locations to require specific progression items belonging to one of those other AHiT worlds. I added some extra print statements to make sure that it would sometimes check if the sweep was finished (by checking every player), but then find newly reachable locations and then continue sweeping.

I also tried adding additional rules to a few entrances to require specific progression items belonging to a different world, but, unsurprisingly, this does not work well because entrance accessibility is cached per-player and only staled when the world the entrance belongs to collects an item, so the required item belonging to a different world could be collected, but without the entrance accessibility ever being staled to result in checking if the entrance is now accessible. This would often result in an error at the end of playthrough calculation because the end of playthrough calculation builds up locations by sphere, but MultiWorld.can_beat_game(), used to cull the spheres, no longer sweeps by spheres, so the two methods of sweeping could stale the cached entrance accessibility at different times, leading to the per-sphere sweep at the end of playthrough calculation to sometimes not be able to reach every required location.

Values in the table are rounded to 1 decimal place.
Note that generation times with every game included do have a small amount of variance in generation time because maybe 1 or 2 games have non-deterministic generation issues.
These times were recorded based on the main branch at f5218fa (before Kingdom Hearts was merged). As the logic in the main branch changes and is merged into this PR, the same seed may take longer or shorter to generate, so these times may not match the current state of main and this PR.

Generated with python -O .\Generate.py --skip_output --seed 1 with a meta.yaml to disable progression balancing.
Yamls 	Before Patch/s 	After Patch/s 	Decrease in duration
1 of each template YAML (67), averaged over 5 individual generations 	122.5 	74.9 	38.9%
2 of each template YAML (134), averaged over 5 individual generations* 	371.6 	192.0 	48.3%
*generated with --seed 2 because --seed 1 produced two sets of very different results, indicating a non-deterministic generation issue that caused two sets of very different seeds to be generated

Generated with python -O .\Generate.py --seed 1 with a meta.yaml to disable progression balancing, and spoiler: 2 in host.yaml.
Yamls 	Before Patch/s 	Only 'Make better use of region accessibility caching' 	After Patch/s 	Decrease in duration before -> after
1 of each non-rom template YAML (53), averaged over 5 individual generations 	504.5 	436.6 	376.4 	25.4%
Above, but playthrough calculation only 	446.9 	390.4 	335.4 	25.0%
Above, but without playthrough calculation 	57.6 	46.2 	41.0 	28.7%