Core: Speed up CollectionState sweeping (#3812)
Increases the performance of sweeping state in
`CollectionState.sweep_for_events()` and `MultiWorld.can_beat_game`,
without changing the generated output (at least for supported games with
their template yamls).

`CollectionState.sweep_for_events()` takes up a significant amount of
generation time, mostly from filling progression items, and the majority
when not generating a playthrough, so improvements to its performance
can have a noticeable effect on generation duration.

Similarly, `MultiWorld.can_beat_game()` takes up a significant amount of
playthrough calculation time from culling spheres.

This patch reduces the duration of both and refactors them to use the
same sweeping code.

This patch has been split from #3688 to focus only on the sweep itself,
and has been simplified to not require any additions to the existing
world API.

### Make better use of region accessibility caching

Region accessibility is cached per-player and the cache for that player
becomes stale whenever an item is collected or removed for that player.
The original sweeping implementations make use of this by only
collecting items once all reachable locations have been found for all
players.

This patch changes to the sweep to find reachable locations per-player,
collect the items from those locations and then move onto the next
player. This keeps region accessibility cached as before, but has the
advantage that if a player's reachable locations contain an item that
gives a later player access to more locations, it will take fewer
iterations for all reachable locations to be collected from.

### Skip checking location accessibility for locations belonging to a
player which did not receive any advancement in the previous iteration

If no advancement items belonging to a player were collected in the
previous sweep iteration (after the point at which their reachable
locations were found), then it is assumed that player should have zero
newly reachable locations, so checking for their reachable locations can
be skipped.

While almost all worlds only logically depend on only their own
items/locations etc., it is possible for a world to logically depend on
a part of another world. To account for this, once the sweep runs out of
accessible locations, the locations of all players are checked to
confirm that the sweep is indeed finished.

My previous idea to account for worlds logically depending on other
worlds was to require all worlds to specify which worlds they logically
depended on. Performing extra sweeps to confirm will be slower than this
previous idea, but it is simpler and does not require any changes to the
existing world API.