OoT: Reorder validate_world so that faster checks are first (#4864)
States are now created lazily because there is no point creating and
sweeping states if faster checks can fail before those states are even
used.

all_state is the most expensive state to create so checks using it are
last. time_travel_state is the next most expensive and then none_state
is the least expensive. Checks that do not need a new state to begin
with are performed first.