Core: Add optional initial bulk fill
Adds an optional bulk fill to the start of progression and priority
fill, enabled through a host.yaml option.

This bulk fill can greatly speed up generation of large multiworlds, but
typically results in a bias towards earlier locations on average. This
bias varies by the worlds being generated and their options, and can, in
some cases, result in less bias towards earlier locations than the bias
introduced by the existing one_item_per_player fill optimisation.

The initial bulk fill places every item onto the first location that
will accept that item while ignoring reachability. Once all items have
been placed, the bulk fill sweeps a state and then, starting from the
last placed item, un-places each item that resulted in an invalid
placement. Un-placed items are collected into the state used to check if
placements are valid. To increase accuracy, after a percentage of placed
items have been un-placed, the state is re-swept. The re-sweep does not
occur after every un-placed item because doing so is too slow.

The main advantage to the bulk fill is that it is considerably faster
than allowing the normal fill_restrictive implementation to place every
item.

The bulk fill cannot place every item and returns control of the fill
back to fill_restrictive once the bulk fill has placed what it can.

As fill_restrictive nears the end of its fill, every sweep to create a
maximum_exploration_state becomes more expensive because more items have
been placed, and when there is only one player with items remaining to
place, then a sweep is performed for every single item to place.

Because the bulk fill attempts to place all items simultaneously and the
invalid placements that have to be un-placed will typically be a
percentage of each world's item pool, the bulk fill ends up partially
normalizing the length of each world's item pool before returning
control back to fill_restrictive. When there are a few item pools that
are much larger than the other item pools, this reduces generation time
of when control is returned to fill_restrictive.

For example, if ten worlds have 100 progression items and one world has
1000, fill_restrictive would normally place 100 items from every world
and then have 900 more items to place for just one world.

If, in this example, the bulk fill managed to place 50% of the items,
there would be 50 remaining progression items to place for the ten
worlds with 100 progression items, and 500 remaining progression items
to place for the world with 1000 progression items. Once there is only
the one larger world with items still to be placed, there will be 450
items to place one at a time, compared to the 900 items to place one at
a time without the bulk fill.