Core: Add optional initial bulk fill
Adds an optional bulk fill to the start of progression and priority
fill, enabled through a host.yaml option. This bulk fill can greatly
speed up generation, but is known to result in a bias towards earlier
locations, from locations, on average, being 50% of the way through the
multiworld spheres, to being 45% of the way through.

The initial bulk fill places every item onto the first location that
will accept that item while ignoring reachability. Once done it sweeps a
state and then, starting from the last placed item, and un-places each
item that resulted in an invalid placement. Un-placed items are
collected into the state used to check if placements are valid. To
increase accuracy, after a percentage of placed items have been
un-placed, the state is re-swept. The re-sweep does not occur after
every un-placed item because doing so is far too slow.

The bulk fill is known to incur a bias towards locations ending up in
earlier spheres.
In a generation of 5 worlds, without this bulk fill, the average sphere
of a location was observed to be 50% of the way through the multiworld's
spheres.
In a generation of those same 5 worlds, with this bulk fill, the average
sphere of a location was observed to be 45% of the way through the
multiworld's spheres, a difference of 5% towards earlier spheres.

This bias is not expected to decrease as the number of worlds increases,
but the current fill_restrictive implementation introduces some of its
own bias/inaccuracy as the number of players increases due to its
optimization that picks one item per-player to place at a time.

The bulk fill does not incur this same one item per-player
bias/inaccuracy so might result in reducing the overall bias as the
number of worlds increases compared to the current fill_restrictive
implementation.

The main advantage to the bulk fill is that it is considerably faster
than allowing the normal fill_restrictive implementation to place every
item.

The bulk fill cannot place every item and returns control of the fill
back to fill_restrictive once the bulk fill has placed what it can.

Because the bulk fill attempts to place all items simultaneously and the
invalid placements that have to be un-placed will typically be a
percentage of each world's item pool, the bulk fill ends up partially
normalizing the length of each world's item pool before returning
control back to fill_restrictive. When there are only a few item pools
that are much larger than the other item pools, this reduces generation
time of when control is returned to fill_restrictive.

For example, if ten worlds have 100 progression items and one world has
1000, fill_restrictive would normally place 100 items from every world
and then have 900 more items to place for just one world. As
fill_restrictive nears the end of its fill, every sweep to create a
maximum_exploration_state becomes more expensive because more items have
been placed, and when there is only one player with items remaining to
place, then a sweep is performed for every single item.
If, in this example, the bulk fill managed to place 50% of the items,
there would be 50 remaining progression items to place for the ten
worlds with 100 progression items, and 500 remaining progression items
to place for the world with 1000 progression items. Once there is only
the one larger world with items still to be placed, there will be 450
items to place one at a time, compared to the 900 items to place one at
a time without the bulk fill.