Fix leftover from old code in unused get_merges_from_adding_piece
`None` is discarded from `found_groups` now, and this code would have
not worked properly if the group with ID `0` ever got used.

The group with ID `0` does not actually get used because not all the IDs
are needed and `0` would be the last ID used because new IDs are taken
from the end of _unused_ids, which are the largest IDs.

get_merges_from_adding_piece is more relevant to the current jigsaw
implementation rather than the older 0.0.3 version this branch is based
on.