Remove redundant reachable location counting in swap
`prev_state` starts off as a copy of `swap_state` and then `swap_state`
collects `item_to_place`. Collecting an item must never reduce
accessibility (otherwise generation breaks horribly), so it is
guaranteed that `swap_state` will always be able to reach at least as
many locations as `prev_state`, so `new_loc_count >= prev_loc_count` is
always `True`.

As a side effect of this change, this fixes generation of Pokemon
Emerald with locally shuffled Badges/HMs when there are world with
unconnected entrances present in the multiworld e.g. KH1. This is
because this location counting did not respect
`single_player_placement=True` and counted reachable locations across
the entire multiworld.