Make logic use pre-created lists
This avoids re-creating lists in logic functions.

Because of the use of Enum values for item/room names instead of string
literals, it is not possible for Python to automatically convert some of
the lists into tuple literals, so all occurrences of lists containing
static elements were moved into the module globals.