Core: Restrictive fill in batches to reduce state sweeping costs (#3872)
Reduces the duration of progression fill by about 40% on average. For
smaller generations, the size of the reduction can vary.

A large performance cost in fill_restrictive is sweeping to create the
maximum exploration state because it sweeps from base_state every time.

This patch breaks up fill_restrictive to fill in batches of items at a
time. Each batch creates its own base state, and additionally a sweep
state created from the batch's base state. The advantage to this is that
the maximum exploration state can be created from the batch's sweep
state rather than having to sweep from base_state.

As a small bonus, the batch's sweep state can sometimes also be used to
create a swap_state, increasing swap performance. A successful swap will
sometimes destroy the batch's sweep state and require it to be
re-created, but because swap sweeps to create a swap_state for every
location it checks, occasionally adding a single extra sweep on a
successful swap is negligible.

There are no expected changes to the contents of the output of
generations with this patch.