Core: Speed up playthrough sphere culling (#3890)
To determine if an item in a sphere is required, playthrough calculation
would remove the item from its location and then use can_beat_game to
sweep through locations until either the game can be beaten (the item is
not required) or until the sweep runs out of reachable locations (the
item is required).

This patch changes how it is determined if an item is required.

The advancement locations in each higher sphere that have already been
culled must be reached for the game to be beatable, so iterate through
the advancement locations in each higher sphere after attempting to cull
an item. If all locations in a higher sphere and all unreachable
locations carried over from a lower sphere are unreachable, then the
game is unbeatable and the item is therefore required to beat the game.

For items which are required to beat the game, this aborts faster than
exhausting all reachable locations to determine that the game is
unbeatable.

For items which are not required to beat the game, the locations which
should be reachable from the current sphere are known in advance, so
this only has to check reachability of the locations in the closest
higher sphere, and any unreachable locations carried over from lower
spheres, before moving onto the next closest sphere, rather than
repeatedly checking all filled locations which have yet to be reached.

This results in a faster playthrough calculation that also scales better
than the previous implementation as the number of advancement items in
the multiworld increases.

Playthrough calculation duration averaged over 5 generations of the same
seed recorded before/s -> after/s.
57 unique games: 671.2 -> 179.5 (73.3% reduction)
28 unique games: 149.2 -> 47.8 (68.0% reduction)
14 unique games: 7.0 -> 4.2 (39.4% reduction)
7 unique games: 4.6 -> 3.4 (26.8% reduction)

Because the spheres are re-arranged and new spheres are added while
iterating, there is no longer a need to calculate the final spheres
because the result is already in its final spheres.

The type hint in create_paths has been changed from List[Set[Location]]
to Collection[Set[Location]] to allow passing it a deque.