Fix reloading the Cantina after entering the junkyard's car
The junkyard area of the Cantina has two cars, one behind a garage door
opened with a lever, and one that has to be built from bricks released
when destroying some trash cans.

When players enter rideable vehicles, they become that vehicle with
their original character set as the riding character of the vehicle
(which is just another character under the hood). Because the car is not
one of the player's unlocked vehicles, this was triggering the check to
reload the cantina.