Replace "All Episodes Token" with "Episode Completion Token"
There is now always a fixed requirement of 6 Tokens.

The number of tokens in the item pool is now knowable in advance because
it is based on the number of enabled chapters. The remaining tokens are
put into starting inventory.

The client maintains backwards compatibility with the older behaviour
when connecting to worlds generated with the older version.