Refactor normal options resolution to avoid option mutability issues
It was too easy for one part of option resolution to modify an option
value and another part of option resolution to read that option value.

Depending on the order of the methods, read option values could become
stale if a later method modified the value.

This change makes it so that option values are read once in advance onto
an object where the attributes can be deleted when they become stale.
Additionally, modifying options now only occurs after option resolution
is completed.

Changes in option values are handled by returning the new value from
methods and passing those values around as arguments to other methods
that need the values.