Remove the core version check on generation
It's irrelevant 99% of the time, it does **not** matter for custom
worlds and YAMLs generated on apgg don't contain that field which means
all the failure cases are basically from custom worlds. And even if
we're talking core worlds, ap core is so slow that we could probably gen
most worlds with YAMLs from 2 years ago and still be fine. (and if not,
options won't match and raise actually useful errors).