Fix goal rarely requiring the gold bricks inside the goal chapter
The goal chapter can require area completions, which can include
bonuses. The problem is that the number of enabled bonuses depends on
how many gold bricks will be logically available, which also includes
the gold bricks in the goal chapter, so the logic could require more
area completions to access the goal chapter than were possible before
gaining access to the goal chapter.

Now, when the combination of goal chapter + goal chapter has gold bricks
+ bonuses are enabled, the number of areas completable prior to the goal
are used when calculating the number of areas to complete to access the
goal.