Fix death counter tracking reading from memory too late
The game updates the death counters after a delay after the area
changes, so the client was getting old values.

The client now sets a huge dummy value, and relies on the
OnGameWatcherTickEvent to set the correct value when it sees that the
real value is lower than the dummy value.

There is a special case for when the death count is 1 and certain cheat
flags are enabled.