Refactor out special handling for create_items goal excluded locations
All the excluded locations are now handled together, simplifying the
code a little.

A fixme todo is added because some reserved characters may be filler, so
the current excluded location requirements of create_items can be too
strict. Generally, it is pretty unlikely to run into issues with more
excluded locations than reserved + required items, so this is not a big
problem.