Refactor TrueJediAndMinikitChecker to initialise from enabled locations
Beforehand, it would initialise from the enabled chapters in slot data,
but this would not be correct anymore due to the Goal Chapter being
included in the set of enabled chapters, but sometimes not having any
locations.

Initialising based on enabled locations is more robust than relying on
specific values from slot data.