Add non power brick Extras as AP items and locations
These are enabled by default.

Due to complexities with unlocked Extras being stored in a bitfield,
if the locations are not randomized, they will still exist in the
multiworld, but will contain their vanilla items, this allows for the
items to be received from AP, e.g. start inventory, event when the
locations are not randomized.